server2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-multipart
httpx
passlib[bcrypt]
neo4j
langchain
langchain-core
langchain-community
langchain-google-genai
google-genai
duckduckgo-search
beautifulsoup4
requests
ddgs
langgraph